final changes

+++ src/app/changepswd/changepswd.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { AccountsService } from '../accounts.service';
+import { Conn } from '../conn';
+import * as CryptoJS from 'crypto-js'
 
 @Component({
   selector: 'app-changepswd',
@@ -14,7 +16,7 @@ export class ChangepswdComponent implements OnInit {
   cnewpass:string;
   msg:string;
 
-  constructor(private chngpassservice:AccountsService,private myrouter:Router) { }
+  constructor(private chngpassservice:AccountsService,private loginsrvobj:AccountsService,private myrouter:Router) { }
 
   ngOnInit(): void {
   }
@@ -23,31 +25,50 @@ export class ChangepswdComponent implements OnInit {
   {
     if(this.newpass == this.cnewpass)
     {
-        var databody={
-          uname:sessionStorage.getItem("username"),
-          cpass:this.currpass,
-          npass:this.newpass
-        }
-
-        this.chngpassservice.changepass(databody).subscribe({
-          
+        this.loginsrvobj.login(sessionStorage.getItem("username")).subscribe({
           next:(res)=>{
-            if(res["nModified"] ==1)
-            {
-              // this.msg="Password Changed Successfully";
-              alert("Password Changed Successfully, Re-login to proceed")
-              sessionStorage.clear();
-              this.myrouter.navigateByUrl("/login");
-            } 
-            else
-            {
-              this.msg="Incorrect Current Password"
-            }
+
+            var decypswd= CryptoJS.AES.decrypt(res[0].pass,Conn.skey ).toString(CryptoJS.enc.Utf8);
+              console.log(decypswd);
+              if(decypswd == this.currpass)
+              {
+                var encypswd= CryptoJS.AES.encrypt(this.newpass, Conn.skey).toString();
+                var databody={
+                      uname:sessionStorage.getItem("username"),
+                      npass:encypswd
+                }
+
+                this.chngpassservice.changepass(databody).subscribe({
+          
+                  next:(res)=>{
+                    if(res["nModified"] ==1)
+                    {
+                      // this.msg="Password Changed Successfully";
+                      alert("Password Changed Successfully, Re-login to proceed")
+                      sessionStorage.clear();
+                      this.myrouter.navigateByUrl("/login");
+                    } 
+                    else
+                    {
+                      this.msg="Incorrect Current Password"
+                    }
+                  },
+                  error:(err)=>{
+                    this.msg=err;
+                  }
+                })
+
+              }
+              else
+              {
+                this.msg="Incorrect Current Password"
+              }
+
           },
           error:(err)=>{
-            this.msg=err;
+              this.msg=err;
           }
-        })
+        })   
       
     }
     else
